requirements.txt
requests
beautifulsoup4
playwright
apscheduler
ollama
python-dotenv
pytz
